yt_stream_endpoint/endpoint/requirements.txt
flask
yt-dlp
Pillow
transformers
torch>=1.13.0; platform_machine == "x86_64"
torchvision
opencv-python==4.5.5.62 
timm
opencv-python-headless==4.5.5.62
openai
ffmpeg-python
git+https://github.com/openai/whisper.git
python-dotenv
psycopg2-binary
scikit-image
gunicorn